# staticfiles/openapi-schema.yml
openapi: 3.0.2
info:
  title: 'Flora API'
  version: '1.0.0'
paths:
  /api/v1/projects/?page=n:
    get:
      operationId: allProjects
      description: All art projects
      parameters:
      - name: page
        in: path
        required: false
        description: page number (only 20 projects on page)
        schema:
          type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProjects'
    post:
      operationId: createProject
      description: Create a new project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: "Some project"
                description:
                  type: string
                  minlength: 1
                  example: "Some description"
                category:
                  type: string
                  format: uuid
                  minLength: 36
                  maxLength: 36
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."
                  description:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."
                  category:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."
  /api/v1/projects/{project_pk}/:
    get:
      operationId: concreteProject
      description: The concrete project
      parameters:
      - name: project_pk
        in: path
        required: true
        description: project pk field value (uuid)
        schema:
          type: string
          format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    put:
      operationId: updateProject
      description: Update an existing project
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                  example: "Some project"
                description:
                  type: string
                  minlength: 1
                  example: "Some description"
                category:
                  type: string
                  format: uuid
                  minLength: 36
                  maxLength: 36
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          content:
            application/json:
              schema:
                type: object
                properties:
                  title:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."
                  description:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."
                  category:
                    type: array
                    items:
                      type: string
                      example: "This field may not be blank."

components:
  schemas:
    Project:
      type: object
      properties:
        pk:
          type: string
          format: uuid
          minLength: 36
          maxLength: 36
        title:
          type: string
          example: "Some title"
          maxLength: 255
          minLength: 1
        description:
          type: string
          example: "Some description"
          minLength: 1
        images:
          type: array
          items:
            type: string
            example: "/media/projects/some_image.png"
            minLength: 1
        category:
          type: object
          properties:
            pk:
              type: string
              format: uuid
              minLength: 36
              maxLength: 36
            title:
              type: string
              example: "Some category"
              minLength: 1
              maxLength: 255
        user:
          type: string
          example: "user"
        pub_datetime:
          type: string
          example: "2021-12-17T19:26:52.517853Z"
    ListProjects:
      type: array
      items:
        $ref: '#/components/schemas/Project'